js library clean up
