+++ src/js/script.js
@@ -26,3 +26,12 @@ $().ready( function () {
 		$('.front-image img').attr('src','/assets/images/app-android.jpg')
 	}
 });
+
+criticalMapsUtils = {
+   convertCoordinateFormat : function ( oldFormat ) {
+    oldFormat = oldFormat.toString();
+    var chars = oldFormat.split( '' );
+    chars.splice( -6, 0, '.' );
+    return chars.join( '' );
+  }
+}
